Remove fake engine
