+++ Etabs_Adapter/CRUD/Create/Load.cs
@@ -43,84 +43,7 @@ namespace BH.Adapter.ETABS
 #endif
     {
         /***************************************************/
-
-        public bool CreateObject(Loadcase loadcase)
-        {
-            eLoadPatternType patternType = loadcase.Nature.ToCSI();
-            double selfWeight = 0;
-
-            int ret = m_model.LoadPatterns.Add(loadcase.Name, patternType, selfWeight, true);
-            loadcase.CustomData[AdapterIdName] = loadcase.Name;
-
-            return true;
-        }
-
-        /***************************************************/
-
-        private bool CreateObject(MassSource massSource)
-        {
-            bool includeElements = massSource.ElementSelfMass;
-            bool includeAddMass = massSource.AdditionalMass;
-            bool includeLoads = massSource.FactoredAdditionalCases.Count > 0;
-
-            int count = massSource.FactoredAdditionalCases.Count;
-            string[] cases = new string[count];
-            double[] factors = new double[count];
-
-            for (int i = 0; i < count; i++)
-            {
-                cases[i] = massSource.FactoredAdditionalCases[i].Item1.Name;
-                factors[i] = massSource.FactoredAdditionalCases[i].Item2;
-            }
-
-            if (m_model.PropMaterial.SetMassSource_1(ref includeElements, ref includeAddMass, ref includeLoads, count, ref cases, ref factors) == 0) { }
-            else
-            {
-                CreateElementError("mass source", massSource.Name);
-            }
-
-            return true;
-        }
-
-        /***************************************************/
-
-        private bool CreateObject(ModalCase modalCase)
-        {
-            return false;
-        }
-
-        /***************************************************/
-
-        private bool CreateObject(LoadCombination loadCombination)
-        {
-            if (m_model.RespCombo.Add(loadCombination.Name, 0) == 0) //0=case, 1=combo 
-            {
-                foreach (var factorCase in loadCombination.LoadCases)
-                {
-                    double factor = factorCase.Item1;
-                    Type lcType = factorCase.Item2.GetType();
-                    string lcName = factorCase.Item2.Name;// factorCase.Item2.Name;// Number.ToString();
-                    eCNameType cTypeName = eCNameType.LoadCase;
-
-                    if (lcType == typeof(Loadcase))
-                        cTypeName = eCNameType.LoadCase;
-                    else if (lcType == typeof(LoadCombination))
-                        cTypeName = eCNameType.LoadCombo;
-
-                    m_model.RespCombo.SetCaseList(loadCombination.Name, ref cTypeName, lcName, factor);
-                }
-                loadCombination.CustomData[AdapterIdName] = loadCombination.Name;
-            }
-            else
-            {
-                CreateElementError(loadCombination.GetType().ToString(), loadCombination.Name);
-            }
-
-            return true;
-        }
-
-        /***************************************************/
-
+        
         private bool CreateObject(ILoad bhLoad)
         {
             SetLoad(bhLoad as dynamic, this.EtabsConfig.ReplaceLoads);
