+++ Etabs_Adapter/CRUD/Read/Load.cs
@@ -46,89 +46,7 @@ namespace BH.Adapter.ETABS
 #endif
     {
         /***************************************************/
-
-        private List<Loadcase> ReadLoadcase(List<string> ids = null)
-        {
-            int nameCount = 0;
-            string[] nameArr = { };
-
-            List<Loadcase> loadcaseList = new List<Loadcase>();
-
-            if (ids == null)
-            {
-                m_model.LoadPatterns.GetNameList(ref nameCount, ref nameArr);
-                ids = nameArr.ToList();
-            }
-
-            foreach (string id in ids)
-            {
-                Loadcase bhLoadcase = new Loadcase();
-
-                eLoadPatternType type = eLoadPatternType.Other;
-
-                if (m_model.LoadPatterns.GetLoadType(id, ref type) == 0)
-                {
-                    bhLoadcase.Name = id;
-                    bhLoadcase.Nature = type.ToBHoM();
-                }
-
-                loadcaseList.Add(bhLoadcase);
-            }
-
-            return loadcaseList;
-        }
-
-        /***************************************************/
-
-        private List<LoadCombination> ReadLoadCombination(List<string> ids = null)
-        {
-            List<LoadCombination> combinations = new List<LoadCombination>();
-
-            //get all load cases before combinations
-            Dictionary<string, Loadcase> bhomCases = ReadLoadcase().ToDictionary(x => x.Name.ToString());
-
-            int nameCount = 0;
-            string[] nameArr = { };
-
-            if (ids == null)
-            {
-                m_model.RespCombo.GetNameList(ref nameCount, ref nameArr);
-                ids = nameArr.ToList();
-            }
-            
-            foreach (string id in ids)
-            {
-                LoadCombination combination = new LoadCombination();
-
-                string[] caseNames = null;
-                double[] factors = null;
-                int caseNum = 0;
-                eCNameType[] nameTypes = null;//<--TODO: maybe need to check if 1? (1=loadcombo)
-
-                if (m_model.RespCombo.GetCaseList(id, ref caseNum, ref nameTypes, ref caseNames, ref factors) == 0)
-                {
-                    combination.Name = id;
-
-                    if (caseNames != null)
-                    {
-                        Loadcase currentCase;
-
-                        for (int i = 0; i < caseNames.Count(); i++)
-                        {
-                            if (bhomCases.TryGetValue(caseNames[i], out currentCase))
-                                combination.LoadCases.Add(new Tuple<double, ICase>(factors[i], currentCase));
-                        }
-                    }
-
-                    combinations.Add(combination);
-                }
-            }
-
-            return combinations;
-        }
-
-        /***************************************************/
-
+        
         private List<ILoad> ReadLoad(Type type, List<string> ids = null)
         {
             List<ILoad> loadList = new List<ILoad>();
