--- Etabs_Adapter/_Engine/Convert/ToCSI.cs
@@ -1,129 +1,0 @@
-/*
- * This file is part of the Buildings and Habitats object Model (BHoM)
- * Copyright (c) 2015 - 2020, the respective contributors. All rights reserved.
- *
- * Each contributor holds copyright over their respective contributions.
- * The project versioning (Git) records all such contribution source information.
- *                                           
- *                                                                              
- * The BHoM is free software: you can redistribute it and/or modify         
- * it under the terms of the GNU Lesser General Public License as published by  
- * the Free Software Foundation, either version 3.0 of the License, or          
- * (at your option) any later version.                                          
- *                                                                              
- * The BHoM is distributed in the hope that it will be useful,              
- * but WITHOUT ANY WARRANTY; without even the implied warranty of               
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                 
- * GNU Lesser General Public License for more details.                          
- *                                                                            
- *                                                                            
- * You should have received a copy of the GNU Lesser General Public License     
- * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
- */
-
-using BH.oM.Structure.Loads;
-using BH.oM.Structure.Constraints;
-using BH.oM.Structure.MaterialFragments;
-using BH.oM.Geometry;
-
-#if Debug17 || Release17
-using ETABSv17;
-#else
-using ETABS2016;
-#endif
-
-// ******************************************************
-// NOTE
-// These Engine methods are improperly put in the Adapter Project
-// as a temporary workaround to the different naming of ETABS dlls (2016, 2017).
-// Any Engine method that does not require a direct reference to the ETABS dlls
-// must be put in the Engine project.
-// ******************************************************
-
-namespace BH.Engine.ETABS
-{
-    public static partial class Convert
-    {
-        /***************************************************/
-
-        public static eLoadPatternType ToCSI(this LoadNature loadNature)
-        {
-            eLoadPatternType loadType;
-            switch (loadNature)
-            {
-                case LoadNature.Dead:
-                    loadType = eLoadPatternType.Dead;
-                    break;
-                case LoadNature.SuperDead:
-                    loadType = eLoadPatternType.SuperDead;
-                    break;
-                case LoadNature.Live:
-                    loadType = eLoadPatternType.Live;
-                    break;
-                case LoadNature.Wind:
-                    loadType = eLoadPatternType.Dead;
-                    break;
-                case LoadNature.Seismic:
-                    loadType = eLoadPatternType.Quake;
-                    break;
-                case LoadNature.Temperature:
-                    loadType = eLoadPatternType.Temperature;
-                    break;
-                case LoadNature.Snow:
-                    loadType = eLoadPatternType.Snow;
-                    break;
-                case LoadNature.Accidental:
-                    loadType = eLoadPatternType.Braking;
-                    break;
-                case LoadNature.Prestress:
-                    loadType = eLoadPatternType.Prestress;
-                    break;
-                case LoadNature.Other:
-                    loadType = eLoadPatternType.Other;
-                    break;
-                default:
-                    loadType = eLoadPatternType.Other;
-                    break;
-            }
-
-            return loadType;
-
-        }
-
-        /***************************************************/
-
-        public static eMatType ToCSI(this MaterialType materialType)
-        {
-            switch (materialType)
-            {
-                case MaterialType.Aluminium:
-                    return eMatType.Aluminum;
-                case MaterialType.Steel:
-                    return eMatType.Steel;
-                case MaterialType.Concrete:
-                    return eMatType.Concrete;
-                case MaterialType.Timber:
-                    Engine.Reflection.Compute.RecordWarning("ETABS does not contain a definition for Timber materials, the material has been set to type 'Other' with 'Orthotropic' directional symmetry");
-                    return eMatType.NoDesign;
-                case MaterialType.Rebar:
-                    return eMatType.Rebar;
-                case MaterialType.Tendon:
-                    return eMatType.Tendon;
-                case MaterialType.Glass:
-                    Engine.Reflection.Compute.RecordWarning("ETABS does not contain a definition for Glass materials, the material has been set to type 'Other'");
-                    return eMatType.NoDesign;
-                case MaterialType.Cable:
-                    Engine.Reflection.Compute.RecordWarning("ETABS does not contain a definition for Cable materials, the material has been set to type 'Steel'");
-                    return eMatType.Steel;
-                default:
-                    Engine.Reflection.Compute.RecordWarning("BHoM material type not found, the material has been set to type 'Other'");
-                    return eMatType.NoDesign;
-            }
-        }
-
-        /***************************************************/
-
-    }
-
-}
-
